+++ docs/src/examples/QasActionsMenu/ExWithLoading.vue
@@ -4,24 +4,6 @@
     <qas-actions-menu v-bind="props" />
 
     <qas-actions-menu v-bind="propsWithoutSplit" />
-
-    <qas-btn class="q-mt-md" icon="sym_r_person" label="Perfil" loading />
-
-    <br>
-
-    <qas-btn class="q-my-md" icon="sym_r_person" label="Perfil" loading variant="primary" />
-
-    <br>
-
-    <qas-btn icon="sym_r_person" label="Perfil" loading variant="secondary" />
-
-    <br>
-
-    <qas-btn class="q-mt-md" icon="sym_r_person" label="Perfil" variant="secondary" />
-
-    <br>
-
-    <qas-btn class="q-mt-md" icon="sym_r_person" :loading="loading.person" variant="secondary" @click="handleLoading('person')" />
   </div>
 </template>
 
@@ -73,11 +55,11 @@ const props = computed(() => {
   }
 })
 
-// const propsWithoutSplit = computed(() => {
-//   return {
-//     list: list.value
-//   }
-// })
+const propsWithoutSplit = computed(() => {
+  return {
+    list: list.value
+  }
+})
 
 // functions
 async function handleLoading (fieldKey) {